Started adding support for editing pages and started refactoring the current post methods
