--- lib/services/post_services/post_pull_request_editing_service.rb
@@ -1,33 +1,0 @@
-# frozen_string_literal: true
-
-require_relative './base_post_service'
-require_relative '../github_service'
-
-module Services
-  ##
-  # This class is responsible for editing posts that are in PR on a Jekyll website
-  class PostPullRequestEditingService < BasePostService
-    def initialize(repo_name, access_token)
-      super(repo_name, access_token)
-    end
-
-    ##
-    # This method submits changes to a post that is already in PR, commiting and pushing the markdown changes
-    # and any added photos to the branch. Since the post is in PR these changes will be a PR updated to the given branch
-    #
-    # Params:
-    # +post_markdown+::the modified markdown to submit
-    # +post_title+::the title for the existing post
-    # +existing_post_file_path+::the file path to the existing post on GitHub
-    # +ref+::the ref to update
-    def edit_post_in_pr(post_markdown, post_title, existing_post_file_path, ref)
-      ref_name = @github_service.get_ref_name_by_sha(ref)
-      sha_base_tree = @github_service.get_base_tree_for_branch(ref)
-
-      new_tree_sha = create_new_tree(post_markdown, post_title, existing_post_file_path, sha_base_tree)
-      @github_service.commit_and_push_to_repo("Edited post #{post_title}", new_tree_sha, ref, ref_name)
-
-      PostImageManager.instance.clear
-    end
-  end
-end
