+++ tests/services/github_service_test.rb
@@ -10,191 +10,6 @@ class GithubServiceTest < BaseGemTest
     @github_service = Services::GithubService.new(@repo_name, 'auth_token')
   end
 
-  def test_get_all_posts_should_return_all_posts_from_the_jekyll_website
-    # Arrange
-    post1 = create_dummy_api_resource(path: '_posts/post1.md')
-    post2 = create_dummy_api_resource(path: '_posts/post2.md')
-    post3 = create_dummy_api_resource(path: '_posts/post3.md')
-
-    post1_content = create_dummy_api_resource(path: '_posts/post1.md', content: 'post 1 base 64 content')
-    post2_content = create_dummy_api_resource(path: '_posts/post2.md', content: 'post 2 base 64 content')
-    post3_content = create_dummy_api_resource(path: '_posts/post3.md', content: 'post 3 base 64 content')
-    image1_content = create_dummy_api_resource(content: 'imagecontents1', path: 'My File1.jpg')
-    image2_content = create_dummy_api_resource(content: 'imagecontents2', path: 'My File2.jpg')
-
-    post1_markdown = "#post1\r\n![My Alt Text](/assets/img/My File1.jpg)\r\n![My Alt Text](/assets/img/My File2.jpg)"
-
-    post1_model = create_post_model(title: 'post 1', author: 'Andy Wojciechowski', hero: 'hero 1',
-                                    overlay: 'overlay 1', contents: post1_markdown, tags: %w[announcement info])
-    post2_model = create_post_model(title: 'post 2', author: 'Grace Fleming', hero: 'hero 2',
-                                    overlay: 'overlay 2', contents: '##post2', tags: ['announcement'])
-    post3_model = create_post_model(title: 'post 3', author: 'Sabrina Stangler', hero: 'hero 3',
-                                    overlay: 'overlay 3', contents: '###post3', tags: ['info'])
-
-    Services::KramdownService.any_instance.expects(:get_all_image_paths).with(post1_markdown)
-                             .returns(['assets/img/My File1.jpg', 'assets/img/My File2.jpg'])
-    Services::KramdownService.any_instance.expects(:get_all_image_paths).with('##post2').returns([])
-    Services::KramdownService.any_instance.expects(:get_all_image_paths).with('###post3').returns([])
-
-    Octokit::Client.any_instance.expects(:contents)
-                   .with(@repo_name, path: '_posts')
-                   .returns([post1, post2, post3])
-    Octokit::Client.any_instance.expects(:contents)
-                   .with(@repo_name, path: '_posts/post1.md')
-                   .returns(post1_content)
-    Octokit::Client.any_instance.expects(:contents)
-                   .with(@repo_name, path: '_posts/post2.md')
-                   .returns(post2_content)
-    Octokit::Client.any_instance.expects(:contents)
-                   .with(@repo_name, path: '_posts/post3.md')
-                   .returns(post3_content)
-    Octokit::Client.any_instance.expects(:contents)
-                   .with(@repo_name, path: 'assets/img/My File1.jpg')
-                   .returns(image1_content)
-    Octokit::Client.any_instance.expects(:contents)
-                   .with(@repo_name, path: 'assets/img/My File2.jpg')
-                   .returns(image2_content)
-
-    Base64.expects(:decode64).with('post 1 base 64 content').returns('post 1 text content')
-    Base64.expects(:decode64).with('post 2 base 64 content').returns('post 2 text content')
-    Base64.expects(:decode64).with('post 3 base 64 content').returns('post 3 text content')
-
-    Factories::PostFactory.any_instance.expects(:create_post)
-                          .with('post 1 text content', '_posts/post1.md', nil).returns(post1_model)
-    Factories::PostFactory.any_instance.expects(:create_post)
-                          .with('post 2 text content', '_posts/post2.md', nil).returns(post2_model)
-    Factories::PostFactory.any_instance.expects(:create_post)
-                          .with('post 3 text content', '_posts/post3.md', nil).returns(post3_model)
-
-    # Act
-    result = @github_service.get_all_posts
-
-    # Assert
-    assert_equal [post1_model, post2_model, post3_model], result
-
-    assert_equal 2, post1_model.images.length
-    assert_post_image('assets/img/My File1.jpg', 'imagecontents1', post1_model.images[0])
-    assert_post_image('assets/img/My File2.jpg', 'imagecontents2', post1_model.images[1])
-
-    assert_equal 0, post3_model.images.length
-  end
-
-  def test_get_all_posts_in_pr_should_return_all_posts_in_pr
-    # Arrange
-    pr_body = 'This pull request was opened automatically by the SG website editor.'
-    post_content = create_dummy_api_resource(content: 'PR base 64 content', path: 'sample.md')
-    image_content = create_dummy_api_resource(content: 'imagecontents', path: 'sample.jpeg')
-    post_model = create_post_model(title: 'post', author: 'Andy Wojciechowski', hero: 'hero',
-                                   overlay: 'overlay', contents: '#post', tags: %w[announcement info])
-
-    Octokit::Client.any_instance.expects(:pull_requests).with(@repo_name, state: 'open')
-                   .returns([create_pull_request_hash('andy-wojciechowski', 'My Pull Request Body', 2),
-                             create_pull_request_hash('andy-wojciechowski',
-                                                      pr_body, 3)])
-
-    Octokit::Client.any_instance.expects(:pull_request_files).with(@repo_name, 1)
-                   .returns([]).never
-    Octokit::Client.any_instance.expects(:pull_request_files).with(@repo_name, 2)
-                   .returns([]).never
-    Octokit::Client.any_instance.expects(:pull_request_files).with(@repo_name, 3).returns([
-                                                                                            create_pull_request_file_hash('myref', 'sample.md'),
-                                                                                            create_pull_request_file_hash('myref', 'sample.jpeg')
-                                                                                          ])
-
-    Octokit::Client.any_instance.expects(:contents)
-                   .with(@repo_name, path: 'sample.md', ref: 'myref')
-                   .returns(post_content)
-
-    Octokit::Client.any_instance.expects(:contents)
-                   .with(@repo_name, path: 'sample.jpeg', ref: 'myref')
-                   .returns(image_content)
-
-    Base64.expects(:decode64).with('PR base 64 content').returns('PR content')
-    Factories::PostFactory.any_instance.expects(:create_post)
-                          .with('PR content', 'sample.md', 'myref').returns(post_model)
-
-    # Act
-    result = @github_service.get_all_posts_in_pr(pr_body)
-
-    # Assert
-    assert_equal [post_model], result
-
-    assert_equal 1, post_model.images.length
-    assert_post_image('sample.jpeg', 'imagecontents', post_model.images.first)
-  end
-
-  def test_get_post_by_title_should_return_nil_if_the_post_does_not_exist
-    # Arrange
-    post1_model = create_post_model(title: 'post 1', author: 'Andy Wojciechowski', hero: 'hero 1',
-                                    overlay: 'overlay 1', contents: '#post1', tags: %w[announcement info])
-    post2_model = create_post_model(title: 'post 2', author: 'Grace Fleming', hero: 'hero 2',
-                                    overlay: 'overlay 2', contents: '##post2', tags: ['announcement'])
-    post3_model = create_post_model(title: 'post 3', author: 'Sabrina Stangler', hero: 'hero 3',
-                                    overlay: 'overlay 3', contents: '###post3', tags: ['info'])
-
-    @github_service.expects(:get_all_posts).returns([post1_model, post2_model, post3_model])
-
-    # Act
-    result = @github_service.get_post_by_title('a very fake post', nil)
-
-    # Assert
-    assert_nil result
-  end
-
-  def test_get_post_by_title_should_return_a_given_post_by_its_title
-    # Arrange
-    post1_model = create_post_model(title: 'post 1', author: 'Andy Wojciechowski', hero: 'hero 1',
-                                    overlay: 'overlay 1', contents: '#post1', tags: %w[announcement info])
-    post2_model = create_post_model(title: 'post 2', author: 'Grace Fleming', hero: 'hero 2',
-                                    overlay: 'overlay 2', contents: '##post2', tags: ['announcement'])
-    post3_model = create_post_model(title: 'post 3', author: 'Sabrina Stangler', hero: 'hero 3',
-                                    overlay: 'overlay 3', contents: '###post3', tags: ['info'])
-
-    @github_service.expects(:get_all_posts).returns([post1_model, post2_model, post3_model])
-
-    # Act
-    result = @github_service.get_post_by_title('post 2', nil)
-
-    # Assert
-    assert_equal post2_model, result
-  end
-
-  def test_get_post_by_title_should_return_nil_if_the_post_does_not_exist_on_a_given_ref
-    # Arrange
-    post1_model = create_post_model(title: 'post 1', author: 'Andy Wojciechowski', hero: 'hero 1',
-                                    overlay: 'overlay 1', contents: '#post1', tags: %w[announcement info])
-    post2_model = create_post_model(title: 'post 2', author: 'Grace Fleming', hero: 'hero 2',
-                                    overlay: 'overlay 2', contents: '##post2', tags: ['announcement'])
-    post3_model = create_post_model(title: 'post 3', author: 'Sabrina Stangler', hero: 'hero 3',
-                                    overlay: 'overlay 3', contents: '###post3', tags: ['info'])
-
-    @github_service.expects(:get_all_posts_in_pr).returns([post1_model, post2_model, post3_model])
-
-    # Act
-    result = @github_service.get_post_by_title('a very fake post', 'ref')
-
-    # Assert
-    assert_nil result
-  end
-
-  def test_get_post_by_title_should_return_a_given_post_by_its_title_given_a_ref
-    # Arrange
-    post1_model = create_post_model(title: 'post 1', author: 'Andy Wojciechowski', hero: 'hero 1',
-                                    overlay: 'overlay 1', contents: '#post1', tags: %w[announcement info])
-    post2_model = create_post_model(title: 'post 2', author: 'Grace Fleming', hero: 'hero 2',
-                                    overlay: 'overlay 2', contents: '##post2', tags: ['announcement'])
-    post3_model = create_post_model(title: 'post 3', author: 'Sabrina Stangler', hero: 'hero 3',
-                                    overlay: 'overlay 3', contents: '###post3', tags: ['info'])
-
-    @github_service.expects(:get_all_posts_in_pr).returns([post1_model, post2_model, post3_model])
-
-    # Act
-    result = @github_service.get_post_by_title('post 2', 'ref')
-
-    # Assert
-    assert_equal post2_model, result
-  end
-
   def test_get_master_head_sha_should_return_the_sha_of_the_head_of_master
     # Arrange
     Octokit::Client.any_instance.expects(:ref).with(@repo_name, 'heads/master')
